abstracted out the publish button
